+++ packages/playground/Samples/textinput.tsx
@@ -42,7 +42,7 @@ export default class Bootstrap extends React.Component<{}, any> {
   };
 
   onPressShowPassword = () => {
-    var previousState = this.state.passwordHidden;
+    const previousState = this.state.passwordHidden;
     this.setState({passwordHidden: !previousState});
   };