+++ vnext/JSI/Shared/ChakraRuntime.cpp
@@ -711,11 +711,18 @@ JsValueRef CALLBACK ChakraRuntime::HostFunctionCall(
       return RunInMethodContext("HostObject::get", [&]() {
         return chakraRuntime->ToJsValueRef(hostObject->get(*chakraRuntime, propertyId));
       });
-    } else if (
-        GetValueType(propertyName) == JsValueType::JsSymbol &&
-        GetPropertyIdFromSymbol(propertyName) == chakraRuntime->m_propertyId.hostObjectSymbol) {
-      // The special property to retrieve the target object.
-      return target;
+    } else if (GetValueType(propertyName) == JsValueType::JsSymbol) {
+      const auto chakraPropertyId = GetPropertyIdFromSymbol(propertyName);
+      if (chakraPropertyId == chakraRuntime->m_propertyId.hostObjectSymbol) {
+        // The special property to retrieve the target object.
+        return target;
+      } else {
+        auto const &hostObject = *static_cast<std::shared_ptr<facebook::jsi::HostObject> *>(GetExternalData(target));
+        const PropNameIDView propertyId{chakraPropertyId};
+        return RunInMethodContext("HostObject::get", [&]() {
+          return chakraRuntime->ToJsValueRef(hostObject->get(*chakraRuntime, propertyId));
+        });
+      }
     }
 
     return static_cast<JsValueRef>(chakraRuntime->m_undefinedValue);