+++ vnext/Microsoft.ReactNative/Views/TouchEventHandler.cpp
@@ -20,8 +20,8 @@
 #include <winrt/Windows.UI.Core.h>
 #include <winrt/Windows.UI.Input.h>
 
-#ifdef WINUI3_PREVIEW3
-#include <winrt/Microsoft.UI.Input2.Experimental.h>
+#ifdef USE_WINUI3
+#include <winrt/Microsoft.UI.Input.Experimental.h>
 #endif
 
 namespace Microsoft::ReactNative {
@@ -200,7 +200,7 @@ TouchEventHandler::ReactPointer TouchEventHandler::CreateReactPointer(
   pointer.target = tag;
   pointer.identifier = m_touchId++;
   pointer.pointerId = point.PointerId();
-#ifndef WINUI3_PREVIEW3
+#ifndef USE_WINUI3
   pointer.deviceType = point.PointerDevice().PointerDeviceType();
 #else
   pointer.deviceType = point.PointerDeviceType();
@@ -209,9 +209,7 @@ TouchEventHandler::ReactPointer TouchEventHandler::CreateReactPointer(
   pointer.isRightButton = props.IsRightButtonPressed();
   pointer.isMiddleButton = props.IsMiddleButtonPressed();
   pointer.isHorizontalScrollWheel = props.IsHorizontalMouseWheel();
-#ifndef WINUI3_PREVIEW3
   pointer.isEraser = props.IsEraser();
-#endif
 
   UpdateReactPointer(pointer, args, sourceElement);
 
@@ -230,9 +228,7 @@ void TouchEventHandler::UpdateReactPointer(
   pointer.positionRoot = rootPoint.Position();
   pointer.positionView = point.Position();
   pointer.timestamp = point.Timestamp() / 1000; // us -> ms
-#ifndef WINUI3_PREVIEW3
   pointer.pressure = props.Pressure();
-#endif
   pointer.isBarrelButton = props.IsBarrelButtonPressed();
   pointer.shiftKey = 0 != (keyModifiers & static_cast<uint32_t>(winrt::Windows::System::VirtualKeyModifiers::Shift));
   pointer.ctrlKey = 0 != (keyModifiers & static_cast<uint32_t>(winrt::Windows::System::VirtualKeyModifiers::Control));