+++ vnext/Scripts/rnw-dependencies.ps1
@@ -58,7 +58,7 @@ function CheckVS {
     if (!(Test-Path $vsWhere)) {
         return $false;
     }
-    $output = & $vsWhere -version 16 -requires $vsComponents -property productPath
+    $output = & $vsWhere -version 16.5 -requires $vsComponents -property productPath
 
     return ($output -ne $null) -and (Test-Path $output);
 }
@@ -67,8 +67,8 @@ function InstallVS {
     $installerPath = "${env:ProgramFiles(x86)}\Microsoft Visual Studio\Installer";
     $vsWhere = "$installerPath\vswhere.exe"
     if (Test-Path $vsWhere) {
-        $channelId = & $vsWhere -version 16 -property channelId
-        $productId = & $vsWhere -version 16 -property productId
+        $channelId = & $vsWhere -version 16.5 -property channelId
+        $productId = & $vsWhere -version 16.5 -property productId
     }
 
     if (!(Test-Path $vsWhere) -or ($channelId -eq $null) -or ($productId -eq $null)) {
@@ -79,8 +79,8 @@ function InstallVS {
         } else {
             & choco install -y visualstudio2019community
         }
-        $channelId = & $vsWhere -version 16 -property channelId
-        $productId = & $vsWhere -version 16 -property productId
+        $channelId = & $vsWhere -version 16.5 -property channelId
+        $productId = & $vsWhere -version 16.5 -property productId
     }
 
     $vsInstaller = "$installerPath\vs_installer.exe"
@@ -172,7 +172,7 @@ $requirements = @(
         Install = { choco install -y git };
     },
     @{
-        Name = 'VS 2019 with UWP and Desktop/C++';
+        Name = 'Visual Studio >= 16.5 with UWP and Desktop/C++';
         Tags = @('appDev', 'vs2019');
         Valid = CheckVS;
         Install = { InstallVS };