+++ vnext/Shared/InstanceManager.h
@@ -40,6 +40,7 @@ struct InstanceWrapper {
 // Things that used to be exported from InstanceManager, but probably belong
 // elsewhere
 std::shared_ptr<InstanceWrapper> CreateReactInstance(
+    std::shared_ptr<Instance> &&instance,
     std::string &&jsBundleRelativePath,
     std::vector<
         std::tuple<std::string, facebook::xplat::module::CxxModule::Provider, std::shared_ptr<MessageQueueThread>>>
@@ -51,6 +52,7 @@ std::shared_ptr<InstanceWrapper> CreateReactInstance(
     std::shared_ptr<DevSettings> devSettings) noexcept;
 
 std::shared_ptr<InstanceWrapper> CreateReactInstance(
+    std::shared_ptr<Instance> &&instance,
     std::string &&jsBundleBasePath,
     std::string &&jsBundleRelativePath,
     std::vector<