+++ packages/@rnw-scripts/eslint-config/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @rnw-scripts/eslint-config
 
-This log was last generated on Thu, 01 Oct 2020 05:08:11 GMT and should not be manually modified.
+This log was last generated on Tue, 24 Nov 2020 05:08:17 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.6
+
+Tue, 24 Nov 2020 05:08:17 GMT
+
+### Patches
+
+- Add more eslint rules (ngerlem@microsoft.com)
+
+## 0.1.5
+
+Fri, 20 Nov 2020 05:08:43 GMT
+
+### Patches
+
+- Integrate 11/18 RN Nightly Build 🎉 (ngerlem@microsoft.com)
 
 ## 0.1.4