+++ packages/react-native-windows-codegen/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - react-native-windows-codegen
 
-This log was last generated on Sat, 07 Nov 2020 05:04:58 GMT and should not be manually modified.
+This log was last generated on Tue, 24 Nov 2020 05:08:17 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.10
+
+Tue, 24 Nov 2020 05:08:17 GMT
+
+### Patches
+
+- Add more eslint rules (ngerlem@microsoft.com)
 
 ## 0.1.8